+++ src/lightly_train/_task_models/ltdetr_object_detection/task_model.py
@@ -18,10 +18,8 @@ from torch import Tensor
 from typing_extensions import Self, override
 
 from lightly_train._export import tensorrt_helpers
-from lightly_train._export.export_onnx import (
-    ONNXExportMixin,
-    ONNXExportPrecisionPolicy,
-)
+from lightly_train._export.export_migraphx import MIGraphXExportMixin
+from lightly_train._export.export_onnx import ONNXExportPrecisionPolicy
 from lightly_train._models import package_helpers
 from lightly_train._models.dinov2_vit.dinov2_vit import DINOv2ViTModelWrapper
 from lightly_train._models.dinov2_vit.dinov2_vit_package import DINOV2_VIT_PACKAGE
@@ -113,7 +111,7 @@ def _resolve_transformer_config(
     return config_factory()
 
 
-class LTDETRObjectDetection(TaskModel, ONNXExportMixin):
+class LTDETRObjectDetection(TaskModel, MIGraphXExportMixin):
     model_suffix = "ltdetr"
 
     def __init__(